+++ lib/ansible/ruby/dsl_builders/task_spec.rb
@@ -228,6 +228,25 @@ describe Ansible::Ruby::DslBuilders::Task do
       end
     end
 
+    context 'splat' do
+          let(:ruby) do
+            <<-RUBY
+            with_items(jinja('servers1'), jinja('servers2')) do |item|
+              copy do
+                src item
+                dest '/file2.conf'
+              end
+            end
+            RUBY
+          end
+
+          it do
+            is_expected.to have_attributes name: 'Copy something',
+                                           with_items: ['{{ servers1 }}', '{{ servers2 }}'],
+                                           module: have_attributes(src: '{{ item }}')
+          end
+        end
+
     context 'free form' do
       before do
         klass = Class.new(Ansible::Ruby::Modules::Base) do